third_party/blink/web_tests/http/tests/loading/simple-subframe-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
frame "f1" - BeginNavigation request to 'data:text/html,<b>test</b>', http method GET
frame "f1" - DidStartNavigation
main frame - didFinishDocumentLoadForFrame
frame "f1" - ReadyToCommitNavigation
frame "f1" - didCommitLoadForFrame
frame "f1" - didFinishDocumentLoadForFrame
frame "f1" - didHandleOnloadEventsForFrame
frame "f1" - didFinishLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a test of load callbacks. It is only useful inside the regression test tool.
